Move LocalFeedLoader and FeedStore collaborator to its own production file
