+++ EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -1,0 +1,44 @@
+//
+//  LocalFeedLoader.swift
+//  EssentialFeed
+//
+//  Created by Gabriel Maldonado Almendra on 14/4/23.
+//
+import Foundation
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
+
+public final class LocalFeedLoader {
+    
+    private let store: FeedStore
+    private let currentDate:  () -> Date
+    
+    public init(store: FeedStore, currentDate: @escaping () -> Date) {
+        self.store = store
+        self.currentDate = currentDate
+    }
+    
+    public func save(items: [FeedItem], completion: @escaping (Error?) -> Void ) {
+        store.deleteCachedFeed { [weak self] error in
+            guard let self = self else { return }
+            if let cacheDeletionError = error {
+                completion(cacheDeletionError)
+            } else {
+                self.cacheItems(items, with: completion)
+            }
+        }
+    }
+    
+    private func cacheItems(_ items: [FeedItem], with completion: @escaping (Error?) -> Void) {
+        store.insert(items, timestamp: currentDate(), completion: { [weak self] error in
+            guard self != nil else { return }
+            completion(error)
+        })
+    }
+}
